+++ frontend/src/components/Instance.css
